수정사항 반영
