--- src/main/java/com/seolandfriends/byeolbyeolcoffee/admin/command/domain/repository/FranchiseRepository.java
@@ -1,9 +1,0 @@
-package com.seolandfriends.byeolbyeolcoffee.admin.command.domain.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import com.seolandfriends.byeolbyeolcoffee.admin.command.domain.aggregate.entity.Franchise;
-@Repository
-public interface FranchiseRepository extends JpaRepository<Franchise, Long> {
-}
