+++ src/main/java/com/seolandfriends/byeolbyeolcoffee/recipe/command/application/dto/RecipeDto.java
@@ -7,6 +7,7 @@ import java.util.List;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.seolandfriends.byeolbyeolcoffee.recipe.command.domain.aggregate.vo.BaseBeverageVO;
 
 import lombok.AllArgsConstructor;
