+++ src/main/java/com/seolandfriends/byeolbyeolcoffee/recipeOption/command/application/dto/CustomOptionDTO.java
@@ -1,0 +1,18 @@
+package com.seolandfriends.byeolbyeolcoffee.recipeOption.command.application.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class CustomOptionDTO {
+	private Long ingredientId;
+
+	private String ingredientName;
+
+	private String ingredientUnit;
+
+	private boolean processing;
+}
