+++ src/main/java/com/seolandfriends/byeolbyeolcoffee/report/command/domain/repository/ReportRepository.java
@@ -1,0 +1,9 @@
+package com.seolandfriends.byeolbyeolcoffee.report.command.domain.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.seolandfriends.byeolbyeolcoffee.report.command.domain.aggregate.entity.Report;
+@Repository
+public interface ReportRepository extends JpaRepository<Report, Long> {
+}
